Lazy-load the admin toolbar for anonymous visitors
The admin toolbar (and the Framer Motion it pulls in) shipped to every visitor and only decided client-side whether to render, so anonymous readers downloaded and executed it just to render nothing. It is now split behind a lightweight gate that dynamically imports the full toolbar only for editor/change-request/revision views and authenticated organization members.
